rename ui reducer file
