+++ src/react/reducers/ui.js
